% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timelines_per_tweeter.R
\name{plot_timelines_per_tweeter}
\alias{plot_timelines_per_tweeter}
\title{plot_timelines_per_tweeter}
\usage{
plot_timelines_per_tweeter(timelines)
}
\arguments{
\item{timelines}{as retrieved by rtweet::get_timelines()}
}
\value{
plot tweet timelines per user
}
\description{
plot_timelines_per_tweeter
}
